requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==2.2.3
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
mysqlclient==2.2.0
email-validator==2.0.0
itsdangerous==2.1.2
pymysql==1.1.0
mysqlclient==2.2.0
